# openapi/saraya.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "Users_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAdministrationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "Users_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMutationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/role": {
      "patch": {
        "operationId": "Users_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMutationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/toggle-status": {
      "patch": {
        "operationId": "Users_toggleStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMutationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/reset-password": {
      "patch": {
        "operationId": "Users_resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/hr/payroll/generate": {
      "post": {
        "operationId": "Payroll_generatePayroll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratePayrollPeriodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تم إنشاء كشوف الرواتب بنجاح"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "HR - Payroll"
        ]
      }
    },
    "/api/v1/hr/payroll/{id}/approve": {
      "post": {
        "operationId": "Payroll_approvePayroll",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "تم اعتماد كشوف الرواتب"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "HR - Payroll"
        ]
      }
    },
    "/api/v1/hr/payroll/{id}": {
      "delete": {
        "operationId": "Payroll_deleteDraft",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "تم حذف مسودة الرواتب"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "HR - Payroll"
        ]
      }
    },
    "/api/v1/hr/payroll/{id}/pay": {
      "post": {
        "operationId": "Payroll_payPayroll",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessPayrollPaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم صرف الرواتب"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "HR - Payroll"
        ]
      }
    },
    "/api/v1/hr/payroll/periods": {
      "get": {
        "operationId": "Payroll_findAllPeriods",
        "parameters": [],
        "responses": {
          "200": {
            "description": "قائمة فترات الرواتب"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "HR - Payroll"
        ]
      }
    },
    "/api/v1/hr/payroll/periods/{id}/slips": {
      "get": {
        "operationId": "Payroll_findSlips",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "تفاصيل القسائم لفترة الراتب"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "HR - Payroll"
        ]
      }
    },
...
  },
  "info": {
    "title": "منظومة السرايا لإدارة مزارع الماشية والألبان والتسمين - SarayaLivestock API",
    "description": "واجهات برمجة التطبيقات المتكاملة لإدارة القطيع، التناسل، الحلب، التسمين، التغذية، والمالية",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Short-lived access token bound to a revocable server session"
      }
    },
    "schemas": {
...
      "UserRole": {
        "type": "string",
        "enum": [
          "SUPER_ADMIN",
          "FARM_MANAGER",
          "VETERINARIAN",
          "MILKER",
          "ACCOUNTANT",
          "WORKER"
        ]
      },
...
      "UserFarmResponseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "UserAdministrationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "farmId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "farm": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserFarmResponseDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "username",
          "fullName",
          "email",
          "role",
          "isActive",
          "farmId",
          "createdAt",
          "farm"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "maxLength": 100
          },
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50
          },
          "email": {
            "type": "string",
            "maxLength": 255
          }
        },
        "required": [
          "fullName",
          "username"
        ]
      },
      "UserMutationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "username",
          "fullName",
          "role",
          "isActive"
        ]
      },
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "FARM_MANAGER",
              "VETERINARIAN",
              "MILKER",
              "ACCOUNTANT",
              "WORKER"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 200
          }
        },
        "required": [
          "password"
        ]
      },
      "ResetPasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "GeneratePayrollPeriodDto": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "description": "تاريخ بداية الفترة"
          },
          "endDate": {
            "type": "string",
            "description": "تاريخ نهاية الفترة"
          },
          "monthName": {
            "type": "string",
            "description": "اسم شهر الراتب"
          }
        },
        "required": [
          "startDate",
          "endDate",
          "monthName"
        ]
      },
      "ProcessPayrollPaymentDto": {
        "type": "object",
        "properties": {
          "paymentAccountId": {
            "type": "string",
            "description": "رقم حساب الدفع (النقدية أو البنك)"
          }
        },
        "required": [
          "paymentAccountId"
        ]
      },
...
    }
  }
}